+++ src/nksip_sipapp.erl
@@ -137,7 +137,7 @@
 
 -export([init/1, get_user_pass/3, authorize/4, route/6, invite/3, reinvite/3, cancel/2, 
          ack/3, bye/3, options/3, register/3, info/3, prack/3, update/3,
-         subscribe/3, notify/3]).
+         subscribe/3, notify/3, message/3]).
 -export([ping_update/3, register_update/3, dialog_update/3, session_update/3,
          subscription_update/3]).
 -export([handle_call/3, handle_cast/2, handle_info/2, terminate/2]).
@@ -571,6 +571,23 @@ subscribe(_ReqId, _From, State) ->
 
 notify(_ReqId, _From, State) ->
     {reply, ok, State}.
+
+
+%% @doc This function is called by NkSIP to process a new incoming MESSAGE
+%% request.
+%%
+%% If you reply a 2xx response like `ok'  or `accepted', it means the message
+%% has been received. Use a 6xx response (like `decline') to refuse it.
+%%
+-spec message(ReqId::nksip_request:id(), From::from(), State::term()) ->
+    call_reply(nksip:sipreply()).
+
+message(_ReqId, _From, State) ->
+
+
+
+    {reply, decline, State}.
+
 
 
 %% @doc Called when a dialog has changed its state.